+++ website/js/example.js
@@ -1,5 +1,5 @@
 
-var loader             = ORBIS.create(updateProgress, complete, 0, 0);
+var loader             = ORBIS.create(updateProgress, animateProgress, complete, 0, 0);
 var button             = findById('launcher');
 var progressBar        = findById('progressBar');
 var progressPercentage = findById('progressPercentage');
@@ -13,18 +13,20 @@ function loadAssets(){
 }
 
 function updateProgress( progress, file ) {
-  //console.log(file);
   progressConsole.innerHTML   += progress + '% / ' + file.name + '<br/>';
-  progressBar.value            = progress;
   progressPercentage.innerHTML = progress + '%';
   progressFile.innerHTML       = file.name;
+}
+
+function animateProgress( percentage ){
+  progressBar.value = percentage;
 }
 
 function complete( logs ) {
   console.log(logs);
   progressFile.innerHTML     = 'loading complete';
   progressConsole.innerHTML += 'loading complete<br/>You can check logs in the console.';
-  button.disabled = false;
+  //button.disabled = false;
 }
 
 function findById( id ) {