+++ RELEASE_NOTES.md
@@ -1,6 +1,14 @@
-Version 0.2.1 (incoming)
+Version 0.3.1 (incoming)
 
  * documentation update
+
+Version 0.3.0 (October 7th 2016)
+
+ * New animation feature :
+  * Added **FrameRat.js** library
+  * Added OnAnimate callback returning animated progress if needed
+ * Callback functions checked
+ * Updated example on the website
 
 Version 0.2.0 (September 21st 2016)
 ------------------------------